Registered global variable checkBaffleFlag; can be used for testing with epics var command.
git-svn-id: http://sbfsvn02/gemini-sw/gem/branches/ioc/scs/mk/baffle_diagsBR1@4147 303ae0b8-5cfe-405b-ba84-a597b98f2cc0

+++ scs-mk-iocApp/src/config.h
@@ -64,6 +64,6 @@ long CADfocusPidControl(struct cadRecord* pcad);
 extern char *periscopeOption[];
 extern char *cenBaffle[];
 extern char *depBaffle[];
-extern long checkBaffleFlag;
+extern int checkBaffleFlag;
 
 #endif
